requirements.txt
torch
torchvision
torch_geometric
numpy
pandas
matplotlib
scikit-learn
Pillow
networkx
tqdm
